Add addDependencyBasedRequires Parameter
To avoid duplication of dependency information from the Maven pom file towards the <requires> list of the rpm-maven-plugin configuration a new Parameter is added: addDependencyBasedRequires.

The idea of this feature is that the rpm-maven-plugin itself inspects the dependencies of the project in which it is executed and adds a Require: [rpm-name] >= [rpm-version] for each found RPM dependency.